Keep SDL3's 64-bit tick values whole
SDL3 widened SDL_GetTicks() from Uint32 to Uint64, so the tick members kept
truncating it. Clang and GCC say nothing; MSVC warns, and warnings are errors,
which is the only reason this surfaced as a Windows build failure rather than a
wrap after 49 days on every platform.

+++ soh/soh/Enhancements/controls/Mapper.cpp
@@ -1156,7 +1156,7 @@ void MapperWindow::EndSession(bool applyResult) {
 }
 
 void MapperWindow::UpdateElement() {
-    const uint32_t now = SDL_GetTicks();
+    const uint64_t now = SDL_GetTicks();
 
     if (now - mLastUserFileCheckTicks >= FILE_WATCH_INTERVAL_MS || mLastUserFileCheckTicks == 0) {
         mLastUserFileCheckTicks = now;
